`similar`: always return a `PencilArray`
This PR changes the behaviour of `similar(u::PencilArray, [T], [dims])`
when the `dims` argument is passed.

First note that we cannot construct a `PencilArray` which is "similar"
to `u` but having different dimensions, since in that case we would need
to generate a new `Pencil` configuration which is not supported.

For this reason, and to make sure that the return type doesn't depend on
the value of `dims`, calling `similar` used to return a non-distributed
array (most often a regular `Array`) whenever `dims` was passed.

Now, we instead make sure that `similar` always returns a `PencilArray`.
This is in particular needed if one wants to combine PencilArray,
StructArrays and DiffEq.jl. So, when `dims` is passed, we now throw an
error if `dims` is different from the size of `u`.